takes randOnCanvas out of food gen
