+++ lib/food-generator.js
@@ -1,17 +1,19 @@
 var Food = require('./food');
 var SpeedBoost = require('./speed-boost');
 var Virus = require('./virus');
+var OpenCoordinates = require('./open-coordinates');
 
 function FoodGenerator(width, height) {
   this.canvasWidth = width;
   this.canvasHeight = height;
+  this.coorGenerator = new OpenCoordinates(width, height);
 }
 
 FoodGenerator.prototype = {
   seedFood: function(players) {
     var food = [];
     for(var i = 0; i < 200; i++) {
-      var foodItem = new Food(this.randOnCanvas(players));
+      var foodItem = new Food(this.coorGenerator.create(players));
       food.push(foodItem);
     }
     return food;
@@ -20,7 +22,7 @@ FoodGenerator.prototype = {
   seedSpeedBoosts: function(players) {
     var speedBoosts = [];
     for(var i = 0; i < 15; i++) {
-      var boost = new SpeedBoost(this.randOnCanvas(players));
+      var boost = new SpeedBoost(this.coorGenerator.create(players));
       speedBoosts.push(boost);
     }
     return speedBoosts;
@@ -29,45 +31,21 @@ FoodGenerator.prototype = {
   seedViruses: function(players) {
     var viruses = [];
     for(var i = 0; i < 8; i++) {
-      viruses.push(new Virus(this.randOnCanvas(players)));
+      viruses.push(new Virus(this.coorGenerator.create(players)));
     }
     return viruses;
   },
 
-  randOnCanvas: function(players) {
-    var newCoords =  { x: Math.floor(Math.random() * (this.canvasWidth - 10) + 5),
-                       y: Math.floor(Math.random() * (this.canvasHeight - 10) + 5) };
-    var isNotInsidePlayer = false;
-    var canWidth = this.canvasWidth;
-    var canHeight = this.canvasHeight;
-
-    function checkPlayer(player, i, players){
-      var xDiff = newCoords.x - player.x;
-      var yDiff = newCoords.y - player.y;
-      var distance = Math.sqrt( xDiff*xDiff + yDiff*yDiff);
-      return distance > (player.mass + 20);
-    }
-
-    while(isNotInsidePlayer === false){
-      isNotInsidePlayer = players.every(checkPlayer);
-      if (isNotInsidePlayer === false){
-        newCoords =  { x: Math.floor(Math.random() * (canWidth - 10) + 5),
-                       y: Math.floor(Math.random() * (canHeight - 10) + 5) };
-      }
-    }
-    return newCoords;
-  },
-
   replaceFood: function(allFood, speedBoosts, players, viruses) {
     while(allFood.length < 200) {
-      var foodItem = new Food(this.randOnCanvas(players));
+      var foodItem = new Food(this.coorGenerator.create(players));
       allFood.push(foodItem);
     }
     while(speedBoosts.length < 15) {
-      speedBoosts.push(new SpeedBoost(this.randOnCanvas(players)));
+      speedBoosts.push(new SpeedBoost(this.coorGenerator.create(players)));
     }
     while(viruses.length < 8) {
-      viruses.push(new Virus(this.randOnCanvas(players)));
+      viruses.push(new Virus(this.coorGenerator.create(players)));
     }
   }
 };
